requirements.txt
Flask==2.3.3
pandas==2.2.2
matplotlib==3.8.4
openpyxl==3.1.2
Pillow==10.3.0
numpy==1.26.4
flask-caching==2.1.0
# For Redis-based caching (production-ready):
redis==5.0.4
gunicorn

